requirements.txt
Pillow==10.0.1
pygame==2.5.0
PyYAML==6.0.1
tk==0.1.0
